authorization/requirements.txt
psycopg2==2.9.9
fastapi==0.45.0
uvicorn==0.23.2
sqlalchemy==2.0.23
python-jose[cryptography]
